Add SetFormatType() for FormatRegistry.
